Wire in persistence metrics for visitor metrics
Add failure metric for failure codes indicating bucket inconsistencies
