+++ storage/src/vespa/storage/distributor/visitormetricsset.h
@@ -1,15 +1,13 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 #pragma once
 
+#include "persistence_operation_metric_set.h"
 #include <vespa/metrics/metrics.h>
 #include <vespa/documentapi/loadtypes/loadtypeset.h>
 
 namespace storage {
 
-struct VisitorMetricSet : public metrics::MetricSet {
-    metrics::DoubleAverageMetric latency;
-    metrics::LongCountMetric failed;
-
+struct VisitorMetricSet : public PersistenceOperationMetricSet {
     VisitorMetricSet(MetricSet* owner = nullptr);
     ~VisitorMetricSet();
